Separet Dom manipulation in module and correct undeclared const
